fix: refactor google service
