+++ .env.template
@@ -116,18 +116,5 @@ CHANNEL_MSG_REPUTATION=
 
 ##
 # @optional {string}
-GOOGLE_API_KEY=
-
-
-##
-# @optional {string}
-GOOGLE_SPREADSHEET_ID=
-
-##
-# @optional {string}
-GOOGLE_POSITIVE_SHEET_NAME=
-
-##
-# @optional {string}
-GOOGLE_NEGATIVE_SHEET_NAME=
+GOOGLE_CREDITS=
