+++ package.json
@@ -29,6 +29,7 @@
     "cyrillic-to-translit-js": "^3.1.0",
     "express": "^4.17.3",
     "fuse.js": "^6.5.3",
+    "google-auth-library": "^8.0.1",
     "googleapis": "^100.0.0",
     "grammy": "^1.7.1",
     "grapheme-splitter": "^1.0.4",
