--- src/bot/google/index.js
@@ -1,5 +1,0 @@
-const googleClient = require('./google');
-
-module.exports = {
-  googleClient,
-};
