+++ src/bot/listeners/test-tensor.listener.js
@@ -4,9 +4,9 @@ const { env } = require('typed-dotenv').config();
 
 const { redisService } = require('../../services/redis.service');
 const { errorHandler } = require('../../utils');
-const { creatorId, trainingChat } = require('../../creator');
+const { trainingChat } = require('../../creator');
 const { getTensorTestResult } = require('../../message');
-const { googleClient } = require('../google');
+const { googleService } = require('../../services/google.service');
 
 const defaultTime = 30;
 const removeTime = 30;
@@ -69,10 +69,10 @@ class TestTensorListener {
       const sheetNegativeName = env.GOOGLE_NEGATIVE_SHEET_NAME;
       switch (state) {
         case 'negatives':
-          return googleClient.appendToSheet(sheetId, sheetNegativeName, word);
+          return googleService.appendToSheet(sheetId, sheetNegativeName, word);
 
         case 'positives':
-          return googleClient.appendToSheet(sheetId, sheetPositiveName, word);
+          return googleService.appendToSheet(sheetId, sheetPositiveName, word);
 
         default:
           throw new Error(`Invalid state: ${state}`);
@@ -326,17 +326,17 @@ class TestTensorListener {
        * */
       ctx.api.config.use(throttler);
 
-      if (ctx.from.id !== creatorId) {
-        if (ctx.chat.type !== 'supergroup') {
-          ctx.reply('В особистих не працюю 😝');
-          return;
-        }
-
-        if (ctx.chat.id !== trainingChat) {
-          ctx.reply('Я працюю тільки в одному супер чаті 😝');
-          return;
-        }
-      }
+      // if (ctx.from.id !== creatorId) {
+      //   if (ctx.chat.type !== 'supergroup') {
+      //     ctx.reply('В особистих не працюю 😝');
+      //     return;
+      //   }
+      //
+      //   if (ctx.chat.id !== trainingChat) {
+      //     ctx.reply('Я працюю тільки в одному супер чаті 😝');
+      //     return;
+      //   }
+      // }
 
       const message = ctx.msg.text || ctx.msg.caption;
